requirements.txt

python-telegram-bot==13.15
gspread==5.7.2
google-auth==2.6.6
